backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
nvidia-riva-client==2.17.0
httpx==0.28.1
pydantic==2.10.4
